Refactor serialization so that binary and json share same interface for descriptors
