+++ include/BLIB/Serialization.hpp
@@ -10,5 +10,7 @@
 #include <BLIB/Serialization/Binary.hpp>
 #include <BLIB/Serialization/Buffers.hpp>
 #include <BLIB/Serialization/JSON.hpp>
+#include <BLIB/Serialization/SerializableField.hpp>
+#include <BLIB/Serialization/SerializableObject.hpp>
 
 #endif
