+++ include/BLIB/Serialization/Binary.hpp
@@ -2,8 +2,6 @@
 #include <BLIB/Serialization/Binary/InputStream.hpp>
 #include <BLIB/Serialization/Binary/OutputFile.hpp>
 #include <BLIB/Serialization/Binary/OutputStream.hpp>
-#include <BLIB/Serialization/Binary/SerializableField.hpp>
-#include <BLIB/Serialization/Binary/SerializableObject.hpp>
 #include <BLIB/Serialization/Binary/Serializer.hpp>
 #include <BLIB/Serialization/Binary/VersionedSerializer.hpp>
 #include <BLIB/Serialization/Buffers.hpp>
