+++ include/BLIB/Serialization/Binary/CMakeLists.txt
@@ -3,8 +3,6 @@ target_sources(BLIB PUBLIC
     InputStream.hpp
     OutputFile.hpp
     OutputStream.hpp
-    SerializableField.hpp
-    SerializableObject.hpp
     Serializer.hpp
     VersionedSerializer.hpp
 )
