+++ include/BLIB/Serialization/CMakeLists.txt
@@ -2,6 +2,8 @@ target_sources(BLIB PUBLIC
     Binary.hpp
     Buffers.hpp
     JSON.hpp
+    SerializableField.hpp
+    SerializableObject.hpp
 )
 
 add_subdirectory(Binary)
