+++ include/BLIB/Serialization/JSON.hpp
@@ -1,4 +1,2 @@
 #include <BLIB/Serialization/JSON/JSON.hpp>
-#include <BLIB/Serialization/JSON/SerializableField.hpp>
-#include <BLIB/Serialization/JSON/SerializableObject.hpp>
 #include <BLIB/Serialization/JSON/Serializer.hpp>
