+++ include/BLIB/Serialization/JSON/CMakeLists.txt
@@ -1,6 +1,4 @@
 target_sources(BLIB PUBLIC
     JSON.hpp
     Serializer.hpp
-    SerializableField.hpp
-    SerializableObject.hpp
 )
