+++ include/BLIB/Serialization/SerializableObject.hpp
@@ -1,6 +1,8 @@
-#ifndef BLIB_SERIALIZATION_JSON_SERIALIZABLEOBJECT_HPP
-#define BLIB_SERIALIZATION_JSON_SERIALIZABLEOBJECT_HPP
+#ifndef BLIB_SERIALIZATION_SERIALIZABLEOBJECT_HPP
+#define BLIB_SERIALIZATION_SERIALIZABLEOBJECT_HPP
 
+#include <BLIB/Serialization/Binary/InputStream.hpp>
+#include <BLIB/Serialization/Binary/OutputStream.hpp>
 #include <BLIB/Serialization/JSON/JSON.hpp>
 #include <BLIB/Util/NonCopyable.hpp>
 #include <unordered_map>
@@ -9,8 +11,6 @@
 namespace bl
 {
 namespace serial
-{
-namespace json
 {
 class SerializableFieldBase;
 
@@ -26,7 +26,7 @@ struct SerializableObjectBase : private util::NonCopyable {
      * @param object The object to read into
      * @return True on success, false on error
      */
-    bool deserialize(const Group& value, void* object) const;
+    bool deserializeJSON(const json::Group& value, void* object) const;
 
     /**
      * @brief Serializes this object and its fields into the returned json data
@@ -34,38 +34,49 @@ struct SerializableObjectBase : private util::NonCopyable {
      * @param object The object to serialize
      * @return Group json data containing this object and its fields
      */
-    Group serialize(const void* object) const;
+    json::Group serializeJSON(const void* object) const;
+
+    /**
+     * @brief Serializes the given object to the given stream
+     *
+     * @param stream The stream to serialize to
+     * @param object The object to serialize
+     * @return True on success, false on error
+     */
+    bool serializeBinary(binary::OutputStream& stream, const void* object) const;
+
+    /**
+     * @brief Deserializes from the stream into the given object
+     *
+     * @param stream The stream to read from
+     * @param object The object to read into
+     * @return True on success, false on error
+     */
+    bool deserializeBinary(binary::InputStream& stream, void* object) const;
+
+    /**
+     * @brief Returns the size of the object when serialized, in bytes
+     *
+     * @param object The object to serialize
+     * @return std::size_t The serialized size of the object, in bytes
+     */
+    std::size_t binarySize(const void* object) const;
+
+    /**
+     * @brief Returns the serializable object descriptor for the given type
+     *
+     * @tparam T The type to get the descriptor for
+     * @return SerializableObjectBase& The object descriptor
+     */
+    template<typename T>
+    static SerializableObjectBase& get();
 
 protected:
-    /**
-     * @brief Struct to indicate that no fields are optional
-     *
-     */
-    struct StrictMode {};
-
-    /**
-     * @brief Struct to indicate that all fields are optional
-     *
-     */
-    struct RelaxedMode {};
-
-    /**
-     * @brief Construct a new Serializable Object Base in strict mode
-     *
-     */
-    SerializableObjectBase(StrictMode&&);
-
-    /**
-     * @brief Construct a new Serializable Object Base in relaxed mode
-     *
-     */
-    SerializableObjectBase(RelaxedMode&&);
+    SerializableObjectBase() = default;
 
 private:
-    SerializableObjectBase() = delete;
-
-    const bool strict;
-    std::unordered_map<std::string, const SerializableFieldBase*> fields;
+    std::unordered_map<std::uint16_t, const SerializableFieldBase*> fieldsBinary;
+    std::unordered_map<std::string, const SerializableFieldBase*> fieldsJson;
 
     friend class SerializableFieldBase;
 };
@@ -91,7 +102,12 @@ class SerializableObject : public SerializableObjectBase {
         "Type does not have object descriptor. Specialize SerializableObject or Serializer.");
 };
 
-} // namespace json
+template<typename T>
+SerializableObjectBase& SerializableObjectBase::get() {
+    static SerializableObject<T> descriptor;
+    return descriptor;
+}
+
 } // namespace serial
 } // namespace bl
