+++ src/Serialization/Binary/CMakeLists.txt
@@ -1,5 +1,4 @@
 target_sources(BLIB PRIVATE
     InputFile.cpp
     OutputFile.cpp
-    SerializableObject.cpp
 )
