+++ src/Serialization/CMakeLists.txt
@@ -1,3 +1,8 @@
+target_sources(BLIB PRIVATE
+    SerializableObject.cpp
+)
+
+
 add_subdirectory(Binary)
 add_subdirectory(Buffers)
 add_subdirectory(JSON)
