+++ src/Serialization/JSON/CMakeLists.txt
@@ -2,5 +2,4 @@ target_sources(BLIB PRIVATE
     JSON.cpp
     JSONLoader.hpp
     JSONLoader.cpp
-    SerializableObject.cpp
 )
