+++ tests/Serialization/Binary/CMakeLists.txt
@@ -1,5 +1,4 @@
 target_sources(BLIB.t PRIVATE
-    SerializableObject.t.cpp
     Serializer.t.cpp
     Streams.t.cpp
     VersionedSerializer.t.cpp
