+++ tests/Serialization/CMakeLists.txt
@@ -1,2 +1,6 @@
+target_sources(BLIB.t PUBLIC
+    SerializableObject.t.cpp
+)
+
 add_subdirectory(Binary)
 add_subdirectory(JSON)
