+++ tests/Serialization/JSON/CMakeLists.txt
@@ -1,5 +1,4 @@
 target_sources(BLIB.t PUBLIC
     JSON.t.cpp
-    SerializableObject.t.cpp
     Serializer.t.cpp
 )
